Wait for indexers when shutting down index
When receiving an `exit` message in the index, we flush the
meta state of all partitions but dont wait until all indexers
are finished.

This can cause memory corruption issues, because the indexers
contain pointers to `measurement` structs that are stored
inside the `partition` classes, which are destructed along with
the index.

This could also lead to data loss, when an indexer handles a batch
after the index was already destructed.