Disable metrics in integration tests
It turns out this branch accidentally fixed a flushing issue in the
accountant. We now see a lot more data available, unfortunately also
system details that contain process IDs and timestamps.

To fix this, we could either always include metrics in queries or simply
disable the accountant. This commit does the latter.