Kill terminator after shutdown timeout
We observed a race condition in unit tests where the terminator
could sometimes "miss" a down message if one of the actors it
tried to kill was a detached actor.

In this case, the timeout would trigger but the terminator was
left alive even though all the actors it tries to terminate are
gone, leading to an indefinite system hang at shutdown.

We explicitly kill the terminator as a workaround, until a more
exhaustive shutdown logic using retries is implemented or the
caf bug involving detached actors is fixed.