Add pkgStatic support to the nix scaffold
This commit moves the nix expression for vast into an overlay.
It also adds a package for zeek's broker library to make it possible
to build zeek-to-vast.

At the smae time, overrides for CAF and Apache Arrow are added to
fix their build with a compiler that only targets static binaries,
and link-time optimization in case of CAF.

Lastly, the shell.nix file has been updated to work with the new
overlay and make use of the CMake changes that were added in the
previous commits.