Unify source configuration
This is essentially a bugfix; the configuration for sources and datagram
sources deviated from the configuration for remotely spawning sources a
while back, since the code was untested and took a different code path
for configuration.  This change makes remotely spawning a source take
the same code path as spawning a local source and thus enables remotely
spawning datagram sources.