Update Flatbuffers versioning scheme
Due to the way a flatbuffer is laid on on disk [1],
it is not possible to access specific fields of a
given type without knowing the type. In particular, to
access the `version()` field of some type, one would
have to know the layout of the type first, which presumably
could change in later versions.

The only part of a flatbuffer that resides at a fixed
offset is the 4-byte file identifier, which can be freely
chosen by the user. So we use this for versioning instead.

[1] https://github.com/dvidelabs/flatcc/blob/master/doc/binary-format.md