Keep strong references in terminator
We used to store weak references in the terminator. This caused issues
when the reference count of the actor to be terminated went to 0. Then
the actor exited prematurely with caf::exit_reason::unreachable.

But the terminator is meant to enforce a sequential shutdown logic, so
this should not happen. The terminator now stores strong references of
the actors that it is supposed to teardown.