Trigger VAST Docker CI when VAST workflow changes
We don't currently run the VAST Docker CI workflow for every PR that
changes VAST simply because it takes too long. The VAST workflow
changing is a strong indication for a change in the build process or
required tooling, so we should have the VAST Docker workflow depend on
it.