Shutdown stalled sources properly
This fixes a very old bug: A stalled source that was unable to generate
messages is called in a loop until a message is generated. While this
may be considered faulty behavior in CAF, we must work around this
assumption of a stream source behaving like a generator. This patch
generates an invalid table slice if the reader was unable to generate
events and calls quit in the reader, and discards the invalid table
slice in the importer.