Improve the detail::overload visitor helper
In very much non-scientific tests, this version of `detail::overload`
compiles slightly faster, and has better diagnostics, because the
visitor no longer is an anonymous struct.

Additionally, the brace-list-init constructor allows for using a
trailing comma, which makes git-diffs leaner, and improves formatting in
a few places.

This change is purely syntactical and does not alter the behavior of the
code.