Resolve taxonomies for vast export and co.
The current implementation unifies the signature of all spawn
functions, so we need to convert the resolved expression back to
a string. This unification seems is done to provide an extension
mechanism to add more components. Maybe we can restructure this
logic by giving the spawn functions access to the response promise.