Allow for additions to table_slice FlatBuffers
This commit moves the current FlatBuffers table for `table_slice` into a
`fbs.table_slice.generic.v0` table specific to the current encoding.

This allows for additional encodings to have a different table layout,
such that the current CAF, Apache Arrow, and MessagePack encodings can
be converted to use custom encoding FlatBuffers tables, and additional
versions of the encodings be added in the future.