Wake up idle sources after one second
This change offers a new solution for dealing with sources that
are unable to produce table slices. The problem is that if a
stream source does not push anything to its downstream when it
is invoked, CAF will not invoke it again.

The old solution was 2-fold:
1. Readers did not consider the timeout as long as not a single
   event was produced.
2. If a reader didn't produce any events anyway, a nullptr would
   be injected into the stream.

Especially (1) is problematic, because it would block the source
actor in the generate function indeterminately. This means the
actor can not respond to any other message and its queue would
fill up.

The new approach keeps the timeout in the readers, but happily
returns without pushing messages downstream if no input was seen.
In that case, the source sends a delayed message to itself, telling
it to wake up and check for new input later.

At the same time, the read timeout is drastically reduced to 20
milliseconds to regain responsiveness for the `vast status` command.