Tie the encoding-specific state to table slice
In an ideal world, all table slices encodings would have all their state
readable from the FlatBuffers tables directly, but that is not currently
the case. This affects the layout—which is yet to be converted to
FlatBuffers—, and the Arrow RecordBatch, which is currently stored as an
Arrow Buffer in the FlatBuffer.

This commit ties the encoding-specific state's lifetime to the lifetime
of the chunk that contans the FlatBuffers table, thus removing
unnecessary allocations at the cost of one additional, internal pointer
in the table slice API.