Embed a segment into the partition
This change aligns segments with partitions by putting the former
into the latter. It is composed of these individual changes:
1. `partition.v1` is introduced, which contains a segment in addition
   to the fields from `v0`.
2. A segment_builder is added to the `active_partition`.
3. After collectin the resulting hits for an expression, the
   evaluator now retrieves the matching events from the embedded
   segment.

In order to enable (2), the segment builder does not own its
underlying flatbufferbuilder any more.

In order to enable (3), the evaluator receives a handle to its
partition actor along with the list of unevaluated predicates.

As a consequence of this change, the count command will only work
if candidate checks are enabled. To remediate this, the evaluator
needs to be changed to relay `rank(hits)` to its client
additionally.

Another consequence is that the eraser has no effect any more, but
it is arguably also not needed any longer.

A couple of follow up changes can be done after this:
- The archive, store, segment_store, eraser and get_command can
  be removed from the codebase.
- id assignment can move from the importer to the partition
- The exporter can be simplified.