Add base class for FlatBuffers root table wrappers
This class is an improvement over repeatedly converting the underlying
chunk in wrapper types like `segment` using `fbs::as_flatbuffer<Root>`.

The constructor does the verification of the chunk's content, and the
type models a semi-regular type that is allowed to hold an invalid
state. Contextual conversion to `bool` allows for checking whether the
wrapped type holds a valid FlatBuffers table of the specified root type.

Derived types must define a constructor accepting a `chunk_ptr&&` as its
first argument, and forward to the constructor of `fbs::table<Derived,
Root>`, or delegate to the parent constructor using `using table::table`
directly.