Add base class for FlatBuffers root table builders
This class serves as a building block for builders for FlatBuffers root
table wrapping types. They are hard to get correct, and this class makes
it easier to write correct builders. The design of this class is
intertwined with the design for the new FlatBuffers root table wrapper
base class.