Silence non-sensical GCC warning
GCC warns that the control reaches the end of a non-void function for
the segment FlatBuffers table visit helper, because it does not see that
all cases in the switch are handled. This happens because flatc adds MIN
and MAX values to the switch. Adding cases for these is not ideal
either, since clang warns when to cases handle the same value. So we
call a [[noreturn]] function in the end that will never actually be
called.