Verify file identifier when constructing a table
The `flatbuffers::Verifier::VerifyBuffer<Root>` function takes an
optional file identifier for verification. This commit moves the
FileIdentifier template paramter from `fbs::table_builder` to the
`fbs::table`, such that it can be used in both classes, and changes the
constructor of `fbs::table` to verify the file identifier as well.