Enforce batch timeout
Under certain conditions, events would only be pushed to the
importer after a long delay. This happens specifically when
starting a new souce.

This is caused by the high initial event request for the stream source.
Internally, CAF repeatedly invokes the generator callback until
all buffers are filled to the desired level, or no new data is
put on the downstream buffer. The initial hint for the generator is
100, which multiplied by the default table slice size means the
reader needs to produce 100000 events before the slices are forwarded
to the node.