Resize bloom filters when finalizing partitions
When creating a new address synopsis structure, it is sized
for the worst-case scenario of every event in the synopsis
having a unique ip address.

When finalizing a synopsis, the exact number of unique ips
is known and the minimal size for the synopsis can be computed,
so that's what we do.