Dont assert on bloom filter parameter choices
The `bloom_filter_parameters` class has no input validation,
so users have to rely on the `evaluate` function to see if a
given combination of parameters is valid.

However, this function can die with an assertion failure if some
parameters were set incorrectly, creating a catch-22 situation.

Additionally, this meant that these checks were not performed
in release builds with assertions disabled, potentially leading
to bloom filters with nonsensical settings.