Address review comments for bloom filter resizing
Use a std::set over a std::vector + bloom_filter combination,
since the latter confused multiple independent reviewers.
This also has the benefit of allowing for stricter checks on
where and how the buffered_address_synopsis is used, and we
don't have to rely on the `type` member being annotated with
stringified bloom filter parameters.

Also, pass the actor handle of the index together with
the persist atom, as opposed to when spawning a partition
actor.