Detect and handle breaking changes in schemas
When changing schemas such that types are no longer a superset of
previous versions of the same type, VAST may end up in a state where it
cannot import events of that type.

This change makes the type registry forget about incompatible versions
of types automatically when a new and incompatible type is registered.