Avoid reallocations in the segment builder
This is an experimental commit that sets the initial buffer size of the
underlying 'FlatBufferBuilder' of the 'segment_builder' to 110% of
'vast.max-segment-size'. The 10% extra is needed because the configured
limit is a soft limit, and because there's still some extra scaffolding
in the FlatBuffers table that is only added when we call
`segment_builder::finish()`.