Make test spec a pattern and case-insensitive
Too many times running tests requires fiddling with the exact test name
in the right casing. This commit enhances -t such that it no longer
looks at the casing of the test name.

Moreover, the test specification is now interpreted as full-blown regex
and can match just a fraction of the tests. For example, `-t zeek` runs
all Zeek tests