Fix potential race condition between evaluator and partition
The partition currently assumes strict ownership of its
indexers, and terminates them when it shuts down.

However, the evaluator did not store a partition handle but
just the handles of the relevant indexers.

This could lead to a race condition between an evaluator and a
partition when an active partition finished persisting right when
a new query was spawned, or when a query took long enough to see
its associated passive partition evicted from the LRU cache.