Store partition synopses in separate files
Reading partition information out of the partition flatbuffer
at startup can incur drastic i/o penalties on systems with
aggressive readahead settings.

To protect against that, we use separate files for the partition
flatbuffers.

Note that the separate meta index files are treated as optional,
so this change should be both backwards- and forwards-compatible.