Fix handling of spawn commands returning caf::none
Some spawn commands return `caf::none` instead of a valid actor handle
when the result set of the actor was found to be empty, and the recent
refactoring of the node left this case unhandled. It is not an error not
to return anything here, and this makes it more explicit.