Move stream flush check into active partition when appropriate
The '-b' flag is mainly used in integration tests to be able to
wait for all imported data to be available for subsequent queries.

It used to be implemented in the index, by checking whether the
stream stage was free of any pending messages. This is brittle,
since while the index might have cleaned out its own stream stage,
the data can still be in a stream between partition and indexer
when the next query arrives.

We slightly improve upon that situation by pushing the check one
step further down the road, ie. into the active partition if it
exists.