Add a 'fuzz-export' binary to run a query against an arbitrary byte string
The binary serves as an example using a string of arbitrary bytes
as an export query in a self-contained binary. This can hopefully
serve as a base for creating a fuzzing test target.

Note that CMake integration, location in the filesystem, etc. are
all preliminary for now and will most likely be changed for the
final version.