Ignore static plugins when specified in config
This commit changes VAST to ignore plugins specified in the `vast.yaml`
when static versions of them are already loaded.

Aside from being the correct thing to do, this fixes an issue with the
integration test suite where plugin integration tests specify their name
in a `vast.yaml` file used for their tests.