Fix start command detection for spdlog
When configuring spdlog we do different things for client and server. We
distinguish between the two by checking whether the current subcommand
we're running is `start`.

However, this returned a false positive for a plugin that had a command
`<plugin> start`. This commit changes to checking whether the full name
of the command we're running is `start`.