Fix dependency summary for interface libraries
The issue is two-fold:
- The INTERFACE_INCLUDE_DIRECTORIES property may contain multiple paths
  and even generator expressions
- The LOCATION propery does not work on INTERFACE libraries.

The solution is to create a stub library and get that libraries
location, while also making use of a deprecated CMake feature
([CMP0026](https://cmake.org/cmake/help/latest/policy/CMP0026.html)) to
get the LOCATION for build targets without in a context where generator
expressions may not be used.

In summary: This fixes the dependency summary when using the bundled
tsl::robin-map.