Factor common functionality in component plugin
It turns out that just having a handle to the NODE is already a valuable
plugin. Such a plugin can spawn an actor and register itself with the
NODE. That's it.

Other plugins, like the analyzer plugin, will then inherit from this
plugin.

The component plugin only gives you the node actor and requires that you
implement the status subscriber.