Make find_package(VAST) a no-op for subprojects
This is a convenience change targeted at plugin developers. We override
`find_package(VAST [...])` to be a no-op in subprojects (e.g., plugins
built alongside VAST) such that they are not required to check whether
the VAST target already exists before calling `find_package`.

The general idea here is that the plugin build scaffolding should be the
same for all possible ways to build a VAST plugin. This removes the last
manual check that plugin developers needed to do.