Relax version requirements for plugins
This commit changes the VAST_REGISTER_PLUGIN macro to take between one
and five arguments, making passing of the version optional. If just the
plugin is passed to the macro, the VAST version is used as the plugin
version. If at least the major version is specified, zeroes are used for
the unspecified parts of the version.