Use proper full install dirs for system config
Before this change, the default build configuration would try to load
schemas from `etc/vast/schema` instead of `/usr/local/etc/vast/schema`
on macOS, and instead of `/etc/vast/schema` on Linux systems if
`VAST_INSTALL_DATADIR` was unset or `CMAKE_INSTALL_DATADIR` set to a
relative path instead of an absolute one (it should be relative).

The solution here is to use the full paths instead.

Additionally, there's no need to make the paths configurable, since that
ever only affected where VAST looked for the config/schemas/plugins but
not where we installed them.