Prefix plugin library output names with vast-plugin-
To avoid confusion between the PCAP plugin and libpcap, which both have
a library file named `libpcap.so`, we now prefix the plugin library
output names with `vast-plugin-`. The PCAP plugin library file is now
named `libvast-plugin-pcap.so`.

This is a breaking change: The output name of plugins changed, so
plugins specified with a full name in `vast.plugins` must be adapted
accordingly.