Remove dependency between directory and plugin name
For plugins built alongside VAST there was an undocumented and since
recent changes unnecessary restriction that plugin name and directory
name containing the plugin's top-level `CMakeLists.txt` file must be the
exact same. This commit removes that restriction, and as an added bonus
allows the CMake variable `VAST_PLUGINS` to be specified using paths
relative to the repository root instead of always requiring absolute
paths.