Convert integer to a concrete type
This replaces the integer type alias with a struct definition.
It is a first step and feasibility analysis for converting the
c++ side of the builtin types, i.e. all members of `vast::data`.

Doing so will protect us from implicit conversions between types
that have similar underlying representations, and makes it possible
to use the same representation for multiple types.