Fetch tags in external plugins CI job
Without this change the CI job for building external plugins failed
because it tried to downlod the wrong VAST version, namely
`2021.04.29-rc3-40` instead of `2021.04.29`.

We don't need to re-releae because the job does not attach any release
artifacts, but this change makes sure it works properly on the next
release.