Hide built-in plugins from version command
We support built-in plugins by using the `VAST_REGISTER_PLUGIN` macro
from within a source file that is part of either the `vast` or the
`libvast` targets.  These plugins are always loaded and cannot be
disabled. This channge hides such plugins from the version output, as
they are really not plugins but simply use the plugin framework to add
functionality.

The idea for this came up when reviewing the transforms functionality,
which has some transformation as such built-in plugins directly as part
of `libvast`.