Fix version mismatch detection
The detection did never run because it was only ever run if the log
level was above the warning threshold, but that is a) a useless
optimization and was b) broken because it was not adapted after the
switch to spdlog.

Additionally, this adds an explicit check for plugin version mismatches.