Fix plugin support for multi-config generators
Before this change, plugins did not properly support multi-config
generators, for which the archive, library, and runtime output
directories contain an additional suffix. That broke the lookup of
plugins inside the build directory.

As a side-effect this fixes the directory lookup for relocatable VAST
binaries when `CMAKE_INSTALL_LIBDIR` was more than one additional
directory inside the specified `CMAKE_INSTALL_PREFIX`.