Dont send dynamic type information to connecting sources
The type registry in VAST knows two different sources of type
information: On the one hand, layouts are described in schema
files that are stored on disk and loaded when VAST is started.
The schema files describe the *input* data format that is
expected by the sources.

On the other hand, the type registry would accumulate the layouts
of all table slices going through the importer. Without transforms,
each of these layouts must have been the input format of some source,
so the type registry would redistribute them to newly registered
sources.

With transforms, this is no longer guaranteed, since both client
and server input transformation might have been applied to the
table slice.

Conceptually, the idea of a source not being able to parse its own
input and relying on information from other, independent sources to
do so is also quite weird.

So from now on, we only forward the statically configured type
information from the schema files to connecting sources.