Improve CI cache usage by re-uploading on cache hits
The regular cache action has one shortfall that is often complained
about: Caches are not re-uploaded when there's a cache hit, which works
against Ccache. With this change, we let Ccache itself control the size
of the cache and always upload it, thus letting us rotate cached object
files in CI more efficiently.