Use clang by default in clang-build-analyzer script
Before this change, the `-DVAST_ENABLE_TIME_TRACE` CMake option was
silently ignored when the current compiler was not set to `clang`.  On
many machines, this led to the paradoxical situation where the
`clang-build-analyzer` script by default runs a full build using no
`clang` and no build analysis.

Instead, we now default the `CC` and `CXX` to `clang` and `clang++`
respectively to still respect custom compilers and compiler wrappers,
and check in our CMake code whether the current compiler supports the
`-ftime-trace` flag rather than only enabling it when the compiler ID
matches `clang`.