Make unix dgram metrics sink connectionless
The previous way of setting up the UDS datagram metrics relied on
a fake connection that made it possible to use the iostream
abstraction. However this fake connection would get destroyed if
the listening end deleted it's socket.

This commit introduces a wrapper around sockaddr_un and the client
fd and uses sendto instead of write.