Merge lists from configuration and CLI
Before this change, we merged configuration from different sources in a
way that instead of merging nested lists, we overwrote them—leading to
plugins specified in a configuration file not to load at all when
plugins were specified on the command line as well. Now, we merge the
lists in order of precedence.