Fix integration tests for external plugins
The bare-mode, which we use for integration tests to avoid side-effects
from global configuration files, disabled the addition of
`vast.plugin-dirs` to the plugin search paths. However, after recent
changes that list also contained the options from both the command line
and the environment variable, which were then ignored with bare-mode
enabled.

Since we do not read configuration files that were not explicitly
specified in bare-mode, it is safe to just read the option
`vast.plugin-dirs` even with bare-mode enabled.