Fix dependency graph for VASTTargets exports set
For the Unix Makefiles generator, our CMake code generated a false
install-time dependency for the VASTTargets export set, causing the
installation of libvast-fbs to trigger a rebuild of all other targets in
the VASTTargets set. This did not occur for other generators like Ninja.
The effects could be seen for example in the static binary CI workflow,
which compiled everything in the VASTTargets export set twice.

Installing the generated FlatBuffers interface library in the same
export set fixes the issue. To make that work, the CMake 3.15+ property
PUBLIC_HEADER must be manually filled with the generated headers, as
adding sources to an interface library is a CMake 3.19+ feature.

I verified manually that this works for both Unix Makefiles and Ninja,
and the effects of this should also immediately be visible in CI.