Fix duplication of lists in client command options
We've recently changed VAST to merge lists in its configuration from all
possible sources rather than overwriting them. However, in the
command-executing `run` function we merged configuration options with
the local actor system's options again in case the command invocation
happened remotely, which was no longer a no-op now that we merge lists,
which in turn led to all list options of client commands being
duplicated. This fixes the issue by not merging lists between client and
server when doing so.