Publish a tenzir/vast-dev Docker image
This is a complete refactoring of our Docker images, and trims the size
of the resulting production image down to just 350MB. We now offer a
`tenzir/vast-dev` variant besides the regular `tenzir/vast` Docker
images that is designed to be used in development contexts, e.g., when
building additional plugins. The individual stages are labelled
`development` and `production` respectively. To build the development
image locally, use `docker build . --target development`.