Use millisecond granularity for the request timeout
In some cases we need to calculate a reduced timeout so nested
requests that that can't be handled fast enough don't cause the
failure to bubble up to the root requester.

A concrete example is the status request: the nested handlers can't
wait the same amount as the ones in the layers above, because the
outer requests are sent earlier. For example in the index, if a
single partition fails to respond, we need to stop waiting earlier
and send back the rest of the data, otherwise the entire "index"
section would be missing from the output of `vast status`.

Changing the type of the default value give us more headroom for
those calculations without requiring extra casting of temporaries.