Use better default error formatter
The result of `caf::to_string(err)` is much more readable than our own
`vast::render(err)`; the latter is a remnant of VAST pre CAF 0.17, and
we should be able to remove it in its entirety. This improves the
situation somewhat by using the better rendering function by default.