Make state structs non-aggregate by adding constructors
This works around a change in behavior of `std::is_constructible<T>`,
which now returns `true` for aggregate types because expressions
like

    new (&state_) State(arg);

are parsed as parenthesized aggregate initialization. This lead to
a bunch of SFINAE-expressions used by CAF to go down the wrong
path and lead to a bunch of half-initialized actor states.