Set name and id in passive partition state at spawn
The fields name and id in the `passive_partition_state` used to
be filled in the unpack method that does a delayed initialization
after requesting the on disk data from the filesystem actor.
If any request hits the `passive_partition` before it receives
its flatbuffer chunk these values held incorret data, which lead
to wrong or misleading information in the log messages.

Since the partition id is available in the spawn function we can
simply use it to initialize those parts right away to avoid the
problem described above.