Use CPM.cmake for dependency version pinning
We experienced some pretty bad behavior with {fmt} 8.x, spdlog 1.9.1 and
clang 12.0.1, and with Homebrew only providing latest the development
experience on macOS suffered a lot. This patch changes {fmt} and spdlog
to be fetched via CPM, unless CPM_USE_LOCAL_PACKAGES is ON or
VAST_DEVELOPER_MODE is OFF (which implies the former).

Additionally, this bumps the minimum requirement for {fmt} to 7.1.3 and
gets rid of some workarounds for compatibility with older versions.