Support native plugins in the static binary
The plugin scaffold has an auto-registration feature, but for static
linking this requires the plugin library to be assembled with
`-whole-archive` (or similar, depending on the linker and platform).
libvast doesn't use that option, so auto-registration fails for the
generated static binary because the symbols are omitted.

The solution here is to extract native plugins from the libvast target
into a separate library that is linked appropriately into the VAST
binary directly, forcing the linker not to omit the symbols required for
the auto-registration.