Fix possible garbage in status command output
This had a slight race condition: We captured a `string_view` to every
component's type, and used that as a key in the status output. However,
when small string optimization is used, the addition/removal of
components caused the string views to be invalidated. This change makes
it so the status handler copies the keys.