Implement sum type access for type
This has turned out to be a bit more complex than initially assumed, and
the new tests failed before I made the changes to the sum type access
specialization. We need to do a double dispatch here: From type index to
sum type token index, and then from that to a dispatch function that
calls the visitor with the correct type.

We could theoretically map type index to dispatch function directly, but
that causes a lot of binary bloat because the mapping between type index
and sum type token index needs to no longer be computed just once in
total, but rather once for every instantiation of apply. So this will
have to do with a double dispatch.