Fix possible desync in pending queries map
This changes the INDEX to monitor EXPORTER actors in order to be able to
remove pending queries when an EXPORTER goes down unexpectedly without
cancelling its running queries. This fixes a msimatch in the pending
queries map in the output of `vast status --detailed`, and also stops
the index from doing further unnecessary work if an export was
unexpectedly killed.