Implement record type transformations and merging
This was a tricky one: `record_type::transform` takes a list of
transformations that map fields at given offsets to other fields, which
allows for overwriting, inserting, and deleting fields from records.

Built on top of it is the new merge operations on record types that
takes two record types and merges them.