Workaround ADL in implicit complex type conversion
Depending on the context, converting a complex concrete type to a type
preferred the conversion to the private base type over the existing
implicit constructor, which is forbidden and caused compiler errors.
This commit moves the state of type and the complex concrete types into
a separate struct they inherit to avoid these issues.

Additionally, this adds some comments to the concept definitions to
clarify why each of the requirements exists.