Introduce an explicit query backlog in the index
Before this change the index would push queries back into its own
message queue when no worker was available. This means we don't
have to skip lots of queries any more whenever the index receives
a new message.

The change also introduces a stricter restriction on the number of
parallel queries. The old version would interleave the work for all
queries, i.e. a worker would switch to a random work package from
the pending queries map when it was done with the one before, but
now a worker gets assigned to a query until that is completely done.