Rewrite the simdjson reader for the new type API
This turned out to be a lot more work than I originally had anticipated.
The simdjson reader operated under the assumption that variant ID of a
concrete data exactly matched the variant ID of a concrete type, which
was mostly a coincidence in the old type system (and failed to work with
legacy alias types), and absolutely broke with the new API.

I have taken the opportunity to write proper dispatch handlers for the
reader and to not make it operate on a flattened layout, but rather on
the record type with its proper structure. They should be much more
self-explanatory now: There exists a function `extract(element, type)`
for all JSON element types that extracts a VAST data according to the
type definition while being a lenient as possible.