Convert libvast to the new type system
This is a massive commit. It takes most occurences of the legacy type
system in libvast and rewrites them to use the new type system instead.
With that come a set of bug fixes and improvements that were necessary
because the new type system is much less flexible to force developers to
make less mistakes when using it, which is especially noticeable in the
active partition, the JSON, Zeek TSV, and CSV readers, and in the
transform steps that inferred types. When reviewing this commit I
recommend looking at all the changes carefully on a file-by-file basis.
There is no obviosuly easier way to look at it.

Notably this commit does not care about plugins, additional tools and
tests yet, so don't ever expect this commit to run when bisecting, but
rather consider the whole pull request as one unit.