Support seeds in generic hash function
One deficiency of hash_append is that it doesn't support runtime
seeding. This commit offers an interface for this use case, yielding the
following two user-facing interfaces:

   hash<H>(x, y, z);
   seeded_hash<H>{a, b, c}(x, y, z);

Here, H is the algorithm, x, y, z objects to hash, and a, b, c seeds
given to the constructor of H. The implementation chooses the optimal
code path to perform oneshot or incremental hashing.