Improve the issue templates
This changes the isuse templates to be much shorter. We now have four
categories of issues:
- Bug Report
- Feature Request
- Story
- Blank

Additionally, new links exist to creating a new Q&A discussion in the
repository for usage questions and to our documentation site. The link
to the security policy for reporting security issues remains unchanged.

This commit also touches the pull request template and removes the
description header. For single-commit PRs GitHub automatically enters
the commit description into the PR template field at the very top, which
we then always had to manually move below the description header. This
removes that step by simply making the description the topmost thing of
any pull request.