Factor bloom filter implementation in view type
It turned out that the lookup performance of the frozen Bloom filter
implementation was 1.2x worse than the mutable version. The reason is
pointer chasing that occurs when deconstructing the opaque flatbuffer
into a typed view.

This commit adds a single view abstraction for Bloom filters that
includes the implementation of the logic. Both mutable and immutable can
use this now.

The view abstraction does not incure a performance penalty in the
mutable version.