Straighten the index actor signature
The internal query handler of the index used to return void but
actually sent an unsafe response with the uuid, candidates and
scheduled partitions triple to the requester. The historical
reason for this was to prevent a race condition which happened when
results started flowing to the requester before it received the
initial response. This problem happens when the OS scheduler
suspends the thread between initiating the worker and returning
the triple.

This commit replaces the old workaround with a cleaner approach:
We now use a response promise and deliver it before initiating the
worker. This is enough to guarantee that the messages will arrive
in the expected order at the requester.

Thanks to that the query handler signature now correctly indicates
that the requester should expect a `<uuid, uint32_t, uint32_t>` as
response.